+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 ## Enhancements
 
-- None
+- Add `optionalFrom` for arrays of `RawRepresentable`s
+  [Keith Smiley](https://github.com/keith)
+  [#125](https://github.com/lyft/mapper/pull/125)
 
 # 7.2.0